+++ src/main/java/com/datapath/checklistukraineapp/dao/service/PermissionDaoService.java
@@ -1,6 +1,6 @@
 package com.datapath.checklistukraineapp.dao.service;
 
-import com.datapath.checklistukraineapp.dao.entity.PermissionEntity;
+import com.datapath.checklistukraineapp.dao.entity.classifier.PermissionEntity;
 import com.datapath.checklistukraineapp.dao.repository.PermissionRepository;
 import com.datapath.checklistukraineapp.exception.PermissionException;
 import lombok.AllArgsConstructor;
@@ -22,7 +22,7 @@ public class PermissionDaoService {
         return repository.findByRole(role).orElseThrow(() -> new PermissionException("Permission with role '" + role + "' not found"));
     }
 
-    public PermissionEntity findById(Long id) {
+    public PermissionEntity findById(Integer id) {
         return repository.findById(id).orElseThrow(() -> new PermissionException("Permission with id '" + id + "' not found"));
     }
 }
